+++ features/step_definitions/timelines_given_steps.rb
@@ -118,3 +118,38 @@ Given /^the following types are enabled for projects of type "(.*?)"$/ do |proje
     project.save
   end
 end
+
+Given (/^there are the following work packages(?: in project "([^"]*)")?:$/) do |project_name, table|
+  project = get_project(project_name)
+  create_work_packages_from_table table, project
+end
+
+def create_work_packages_from_table table, project
+  table.map_headers! { |header| header.underscore.gsub(' ', '_') }
+
+  table.hashes.each do |type_attributes|
+    [ ["author", User],
+      ["responsible", User],
+      ["assigned_to", User],
+      ["type", Type],
+      ["fixed_version", Version],
+      ["priority", IssuePriority],
+      ["status", Status],
+      ["parent", WorkPackage]
+    ].each do |key, const|
+      if type_attributes[key].present?
+        type_attributes[key] = InstanceFinder.find(const, type_attributes[key])
+      else
+        type_attributes.delete(key)
+      end
+    end
+
+    # lookup the type by its name and replace it with the type
+    # if the cast is ommitted, the contents of type_attributes is interpreted as an int
+    unless type_attributes.has_key? :type
+      type_attributes[:type] = Type.where(name: type_attributes[:type].to_s).first
+    end
+
+    FactoryGirl.create(:work_package, type_attributes.merge(:project_id => project.id))
+  end
+end