+++ doc/CHANGELOG.md
@@ -32,7 +32,13 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#1281` I18n.js Not working correctly. Always returns English Translations
 * `#1758` Migrate functional-tests for issues into specs for work package
 * `#1771` Fixed bug: Refactor Types Project Settings into new Tab
+* `#1880` Re-introduce at-time scope
+* `#1881` Re-introduce project planning comparison in controller
+* `#1883` Extend at-time scope for status comparison
+* `#1884` Make status values available over API
+* `#1994` Integrational tests for work packages at_time (API)
 * `#2158` Work Package General Setting
+* `#2173` Adapt client-side to new server behavior
 * `#2306` Migrate issues controller tests
 * `#2307` Change icon of home button in header from OpenProjct icon to house icon
 * `#2310` Add proper indices to work_package
@@ -42,6 +48,7 @@ See doc/COPYRIGHT.rdoc for more details.
 
 * `#1560` WorkPackage/update does not retain some fields when validations fail
 * `#1771` Refactor Types Project Settings into new Tab
+* `#1878` Project Plan Comparison(server-side implementation): api/v2 can now resolve historical data for work_packages
 * `#1929` Too many lines in work package view
 * `#1946` Modal shown within in Modal
 * `#1949` External links within modals do not work